+++ packages/liquidity-sdk/tsup.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts"],
+  dts: true,
+  format: ["esm", "cjs"],
+  treeshake: true,
+});
