+++ packages/liquidity-widget/package.json
@@ -11,6 +11,7 @@
   "main": "./dist/index.global.js",
   "module": "./dist/index.js",
   "devDependencies": {
+    "@goodsdks/liquidity-sdk": "*",
     "@repo/typescript-config": "workspace:*",
     "tsup": "^8.3.5",
     "typescript": "latest"
